Update memory utility imports after guest staging move
Update CE call sites to reference the memory utilities in their guest-staging packages after the ownership move.

This commit is intentionally mechanical. It keeps the broad import and reference fallout separate from the move itself so reviewers can distinguish semantic migration work from package-update fallout.
