+++ substratevm/src/com.oracle.svm.core/src/com/oracle/svm/core/JavaMemoryUtil.java
@@ -25,6 +25,8 @@
 package com.oracle.svm.core;
 
 import com.oracle.svm.core.config.ObjectLayout;
+import com.oracle.svm.guest.staging.core.IntrinsicCandidate;
+import com.oracle.svm.guest.staging.core.UnmanagedMemoryUtil;
 import com.oracle.svm.shared.Uninterruptible;
 import org.graalvm.word.Pointer;
 import org.graalvm.word.UnsignedWord;
