+++ substratevm/src/com.oracle.svm.core/src/com/oracle/svm/core/jvmti/JvmtiCapabilitiesUtil.java
@@ -29,7 +29,7 @@ import org.graalvm.nativeimage.Platforms;
 import org.graalvm.nativeimage.c.struct.SizeOf;
 import org.graalvm.word.Pointer;
 
-import com.oracle.svm.core.UnmanagedMemoryUtil;
+import com.oracle.svm.guest.staging.core.UnmanagedMemoryUtil;
 import com.oracle.svm.core.jvmti.headers.JvmtiCapabilities;
 
 /** Methods related to {@link JvmtiCapabilities}. */
